Add bundle for persisting API implementations.
NB: The JPA enhancer included with Aries only works with Java 6 or earlier classfiles.  Therefore, we explicitly set the source/target compatibility to JDK 6.  To ensure no warnings in the build, we use JDK 6 to build the project (by pointing `JAVA_HOME` to a JDK 6 installation) to ensure the correct boot classpath is used.

However, gradle-bundle-plugin version 0.7 and later use Bnd 3.x, which is built using JDK 7.  Therefore, in this configuration, Bnd cannot be executed.  To overcome this, we downgraded to gradle-bundle-plugin 0.6, which uses Bnd 2.x, which is JDK 6 compatible.