Rename project property `JDK6_HOME` to `local_jdk6Home`.
gradle-bundle-plugin passes all String-valued project properties to bnd for use by the Jar builder.  bnd apparently considers any such property that begins with an uppercase letter to be a bundle header, and thus adds it to the manifest.  This behavior was causing `JDK6_HOME` to appear in all bundle manifests, which was not desired.  The new naming convention ensures this property is excluded from the bundle manifest.